Updated pixel shifting
delta_0/delta_1 are now 0 instead of an array of zeros when pixel shifting is off.
Moved the execution of the pixel shift in the already existing definition for pixel shifting.

Implemented a method that catches points going outside of the image. Displacements of these points are set to NaN after this happens. Before, a point going outside of the image caused an error.